map to own component
